+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.7.0
+
+* Introduce `semantic_embed` endpoint on client and model.
+* Introduce timeout on client
+
 ## 1.6.0
 
 * Introduce AlephAlphaModel as a more convenient alternative to direct usage of AlephAlphaClient